Front update [Transfer commit]

+++ src/components/Navigation/Navigation.jsx
@@ -1,25 +1,64 @@
 import React from 'react';
-import { Menu, Icon } from 'antd';
+import {Menu, Icon, Modal, Button} from 'antd';
 
 class Navigation extends React.Component {
+
+    state = { visible: false };
+
+    showModal = () => {
+        this.setState({
+            visible: true,
+        });
+    };
+
+    handleOk = e => {
+        console.log(e);
+        this.setState({
+            visible: false,
+        });
+    };
+
+    handleCancel = e => {
+        console.log(e);
+        this.setState({
+            visible: false,
+        });
+    };
+
     render() {
         return (
-            <Menu mode="horizontal">
+            <Menu mode="horizontal" defaultSelectedKeys={this.props.selected}>
                 <Menu.Item key="home">
-                    <Icon type="home" />Home
+                    <a href="/home"><Icon type="home" />Home</a>
                 </Menu.Item>
-                <Menu.Item key="association">
-                    <Icon type="team" />Associations
+                <Menu.Item key="associations">
+                    <a href="/associations"><Icon type="team" />Associations</a>
                 </Menu.Item>
                 <Menu.Item key="events">
-                    <Icon type="pushpin" />Events
+                    <a href="/events"><Icon type="pushpin" />Events</a>
                 </Menu.Item>
                 <Menu.Item key="orders">
-                    <Icon type="shopping-cart" />My Orders
+                    <a href="/orders"><Icon type="shopping-cart" />My Orders</a>
                 </Menu.Item>
-                <Menu.Item className="force-align-right" key="logout">
+                <Menu.Item className="force-align-right" key="logout" onClick={this.showModal}>
                     <Icon type="poweroff" />Logout
                 </Menu.Item>
+                <Modal
+                    title="Logout"
+                    visible={this.state.visible}
+                    onOk={this.handleOk}
+                    onCancel={this.handleCancel}
+                    footer={[
+                        <Button key="Cancel" onClick={this.handleCancel}>
+                            Cancel
+                        </Button>,
+                        <Button key="Disconnect" type="primary" onClick={this.handleOk}>
+                            Disconnect
+                        </Button>,
+                    ]}
+                >
+                    <p>Are you sure ?</p>
+                </Modal>
             </Menu>
         );
     }
